requirements.txt
flask
flask-cors
flask-bcrypt
pymongo
requests
python-dotenv
gunicorn